+++ acceptance-tests/src/test/java/tech/pegasys/web3signer/dsl/utils/MetadataFileHelpers.java
@@ -188,46 +188,6 @@ public class MetadataFileHelpers {
     }
   }
 
-  public void createInterlockYamlFileAt(
-      final Path metadataFilePath,
-      final Path knownServersFile,
-      final String keyPath,
-      final KeyType keyType) {
-    // these are default credentials of Interlock on USB Armory
-    final Map<String, String> yaml = new HashMap<>();
-    yaml.put("type", "interlock");
-    yaml.put("interlockUrl", "https://10.0.0.1");
-    yaml.put("knownServersFile", knownServersFile.toString());
-    yaml.put("volume", "armory");
-    yaml.put("password", "usbarmory");
-    yaml.put("keyPath", keyPath);
-    yaml.put("keyType", keyType.name());
-
-    createYamlFile(metadataFilePath, yaml);
-  }
-
-  public void createYubihsmYamlFileAt(
-      final Path metadataFilePath,
-      final String pkcs11ModulePath,
-      final String connectorUrl,
-      final String additionalInitConfig,
-      final short authId,
-      final String password,
-      final int opaqueDataId,
-      final KeyType keyType) {
-    final Map<String, Serializable> yaml = new HashMap<>();
-    yaml.put("type", "yubihsm");
-    yaml.put("pkcs11ModulePath", pkcs11ModulePath);
-    yaml.put("connectorUrl", connectorUrl);
-    yaml.put("additionalInitConfig", additionalInitConfig);
-    yaml.put("authId", authId);
-    yaml.put("password", password);
-    yaml.put("opaqueDataId", opaqueDataId);
-    yaml.put("keyType", keyType.name());
-
-    createYamlFile(metadataFilePath, yaml);
-  }
-
   public void createAwsYamlFileAt(
       final Path metadataFilePath,
       final String awsRegion,