+++ signing/src/main/java/tech/pegasys/web3signer/signing/config/metadata/AbstractArtifactSignerFactory.java
@@ -24,8 +24,6 @@ import tech.pegasys.web3signer.keystorage.hashicorp.config.KeyDefinition;
 import tech.pegasys.web3signer.keystorage.hashicorp.config.TlsOptions;
 import tech.pegasys.web3signer.signing.KeyType;
 import tech.pegasys.web3signer.signing.config.AzureKeyVaultFactory;
-import tech.pegasys.web3signer.signing.config.metadata.interlock.InterlockKeyProvider;
-import tech.pegasys.web3signer.signing.config.metadata.yubihsm.YubiHsmOpaqueDataProvider;
 
 import java.io.FileNotFoundException;
 import java.io.IOException;
@@ -39,20 +37,14 @@ public abstract class AbstractArtifactSignerFactory implements ArtifactSignerFac
 
   final HashicorpConnectionFactory hashicorpConnectionFactory;
   final Path configsDirectory;
-  private final InterlockKeyProvider interlockKeyProvider;
-  private final YubiHsmOpaqueDataProvider yubiHsmOpaqueDataProvider;
   private final AzureKeyVaultFactory azureKeyVaultFactory;
 
   protected AbstractArtifactSignerFactory(
       final HashicorpConnectionFactory hashicorpConnectionFactory,
       final Path configsDirectory,
-      final InterlockKeyProvider interlockKeyProvider,
-      final YubiHsmOpaqueDataProvider yubiHsmOpaqueDataProvider,
       final AzureKeyVaultFactory azureKeyVaultFactory) {
     this.hashicorpConnectionFactory = hashicorpConnectionFactory;
     this.configsDirectory = configsDirectory;
-    this.interlockKeyProvider = interlockKeyProvider;
-    this.yubiHsmOpaqueDataProvider = yubiHsmOpaqueDataProvider;
     this.azureKeyVaultFactory = azureKeyVaultFactory;
   }
 
@@ -92,25 +84,6 @@ public abstract class AbstractArtifactSignerFactory implements ArtifactSignerFac
       return Bytes.fromHexString(secret);
     } catch (final Exception e) {
       throw new SigningMetadataException("Failed to fetch secret from hashicorp vault", e);
-    }
-  }
-
-  protected Bytes extractBytesFromInterlock(final InterlockSigningMetadata metadata) {
-    try {
-      return interlockKeyProvider.fetchKey(metadata);
-    } catch (final RuntimeException e) {
-      throw new SigningMetadataException(
-          "Failed to fetch secret from Interlock: " + e.getMessage(), e);
-    }
-  }
-
-  protected Bytes extractOpaqueDataFromYubiHsm(
-      final YubiHsmSigningMetadata yubiHsmSigningMetadata) {
-    try {
-      return yubiHsmOpaqueDataProvider.fetchOpaqueData(yubiHsmSigningMetadata);
-    } catch (final RuntimeException e) {
-      throw new SigningMetadataException(
-          "Failed to fetch opaque data from YubiHSM: " + e.getMessage(), e);
     }
   }