+++ signing/src/main/java/tech/pegasys/web3signer/signing/config/metadata/SignerOrigin.java
@@ -17,8 +17,6 @@ public enum SignerOrigin {
   HASHICORP,
   AWS,
   GCP,
-  INTERLOCK,
-  YUBI_HSM,
   FILE_KEYSTORE,
   FILE_RAW
 }