+++ signing/src/test/java/tech/pegasys/web3signer/signing/config/metadata/BlsArtifactSignerFactoryTest.java
@@ -32,8 +32,6 @@ import tech.pegasys.web3signer.signing.ArtifactSigner;
 import tech.pegasys.web3signer.signing.BlsArtifactSigner;
 import tech.pegasys.web3signer.signing.KeyType;
 import tech.pegasys.web3signer.signing.config.AzureKeyVaultFactory;
-import tech.pegasys.web3signer.signing.config.metadata.interlock.InterlockKeyProvider;
-import tech.pegasys.web3signer.signing.config.metadata.yubihsm.YubiHsmOpaqueDataProvider;
 
 import java.io.IOException;
 import java.nio.file.Files;
@@ -63,8 +61,6 @@ class BlsArtifactSignerFactoryTest {
   private ArtifactSignerFactory artifactSignerFactory;
 
   private Vertx vertx;
-  private InterlockKeyProvider interlockKeyProvider;
-  private YubiHsmOpaqueDataProvider yubiHsmOpaqueDataProvider;
   private AwsSecretsManagerProvider awsSecretsManagerProvider;
   private AzureKeyVaultFactory azureKeyVaultFactory;
 
@@ -84,8 +80,6 @@ class BlsArtifactSignerFactoryTest {
   @BeforeEach
   void setup() {
     vertx = Vertx.vertx();
-    interlockKeyProvider = new InterlockKeyProvider(vertx);
-    yubiHsmOpaqueDataProvider = new YubiHsmOpaqueDataProvider();
     awsSecretsManagerProvider = new AwsSecretsManagerProvider(100);
     azureKeyVaultFactory = new AzureKeyVaultFactory();
 
@@ -94,8 +88,6 @@ class BlsArtifactSignerFactoryTest {
             configDir,
             new NoOpMetricsSystem(),
             new HashicorpConnectionFactory(),
-            interlockKeyProvider,
-            yubiHsmOpaqueDataProvider,
             awsSecretsManagerProvider,
             (args) -> new BlsArtifactSigner(args.getKeyPair(), args.getOrigin()),
             azureKeyVaultFactory);
@@ -103,8 +95,6 @@ class BlsArtifactSignerFactoryTest {
 
   @AfterEach
   void cleanup() {
-    interlockKeyProvider.close();
-    yubiHsmOpaqueDataProvider.close();
     vertx.close();
   }