+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## Upcoming Release
 ### Breaking Changes
 - Deprecated CLI option `--azure-secrets-tags` has been removed. Use `--azure-tags` instead.
+- USB Armory II/Interlock support has been removed.
+- YubiHSM 2 support has been removed.
 
 ### Features Added
 - Updated Teku libraries to handle BPO configurations